kloud: move koding under the provider folder
