+++ go/src/koding/kites/kloud/provider/koding/bucket.go
