+++ go/src/koding/kites/kloud/provider/koding/checker.go
