+++ go/src/koding/kites/kloud/provider/koding/cleaners.go
