+++ go/src/koding/kites/kloud/provider/koding/cloudinit.go
