+++ go/src/koding/kites/kloud/provider/koding/fetcher.go
