+++ go/src/koding/kites/kloud/provider/koding/helper.go
