+++ go/src/koding/kites/kloud/provider/koding/info.go
