+++ go/src/koding/kites/kloud/provider/koding/locker.go
