+++ go/src/koding/kites/kloud/provider/koding/machine.go
