+++ go/src/koding/kites/kloud/provider/koding/permission.go
