+++ go/src/koding/kites/kloud/provider/koding/queue.go
