+++ go/src/koding/kites/kloud/provider/koding/resize.go
